refactor(capture): split external provider actions
