+++ app/lib/pages/home/home_content.dart
@@ -173,12 +173,11 @@ class HomeContentPageState extends State<HomeContentPage> with AutomaticKeepAliv
       PlatformManager.instance.analytics.phoneMicRecordingStarted();
     }
     if (!context.mounted) return;
-    final topConvoId = (captureProvider.conversationProvider?.conversations ?? []).isNotEmpty
-        ? captureProvider.conversationProvider!.conversations.first.id
-        : null;
     Navigator.push(
       context,
-      MaterialPageRoute(builder: (context) => ConversationCapturingPage(topConversationId: topConvoId)),
+      MaterialPageRoute(
+        builder: (context) => ConversationCapturingPage(topConversationId: captureProvider.topConversationId),
+      ),
     );
   }
